refactor: layout for booking form
