--- client/src/components/booking/DateTime.module.css
@@ -1,30 +1,0 @@
-.datetimeWrapper {
-  max-width: 960px;
-  margin: 4em auto;
-  width: 100%;
-}
-
-h2 {
-  text-align: center;
-}
-
-.bookingForm {
-  display: grid;
-  grid-template-columns: [labels] auto [controls] 1fr;
-  grid-auto-flow: row;
-  grid-gap: 0.8em;
-  padding: 1.2em;
-}
-
-.bookingForm > label {
-  grid-column: labels;
-  grid-row: auto;
-}
-
-.bookingForm > input,
-.bookingForm > button,
-.bookingForm > Datetime {
-  grid-column: controls;
-  grid-row: auto;
-  padding: 1em;
-}
