+++ client/src/components/booking/EventBooking.module.css
@@ -1,0 +1,48 @@
+.eventBookingWrapper {
+  max-width: 960px;
+  margin: 4.5rem auto;
+  width: 100%;
+}
+
+h2,
+h3,
+h4 {
+  text-align: center;
+}
+
+.bookingForm {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-auto-flow: row;
+  grid-gap: 0.8rem;
+  padding: 1.2rem;
+}
+
+.eventTitle {
+  grid-column: span 2;
+}
+
+.bookingForm > button {
+  grid-column: span 2;
+  padding: 0.8rem;
+}
+
+.bookingGrid {
+  display: grid;
+  grid-template-columns: [labels] auto [controls] 1fr;
+  grid-auto-flow: row;
+  grid-gap: 0.5rem;
+  padding: 0.5rem;
+  align-items: center;
+}
+
+.bookingGrid > label {
+  grid-column: labels;
+  grid-row: auto;
+}
+
+.bookingGrid > input {
+  grid-column: controls;
+  grid-row: auto;
+  padding: 1rem;
+}
